No more 'SimpleSavedState.ts' file
