+++ www/UI/Sidebar.tsx
@@ -5,10 +5,10 @@ import { ReactElement, useCallback } from 'react';
 import { st } from 'www/Constants';
 import { SetSearch } from 'www/Globals';
 import { CurrentView, Keys, StrId } from 'www/Shared/CommonTypes';
-import { searchTermState } from 'www/State/SimpleSavedState';
+import { curViewState } from 'www/State/CurrentView';
+import { searchTermState } from 'www/State/SongPlayback';
 import { GetHelperText } from 'www/Utils';
 
-import { curViewState } from 'www/State/CurrentView';
 import '../styles/Sidebar.css';
 
 type ViewEntry = {
